requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-filter==24.3
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.1.1
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
gunicorn==23.0.0
whitenoise==6.8.2
